[pytorch][ci] add custom selective build flow for android build
Summary:
Mobile custom selective build has already been covered by `test/mobile/custom_build/build.sh`.
It builds a CLI binary with host-toolchain and runs on host machine to
check correctness of the result.

But that custom build test doesn't cover the android/gradle build part.
And we cannot use it to measure and track the in-APK size of custom
build library.

So this PR adds the selective build test coverage for android NDK build.
Also integrate with the CI to upload the custom build size to scuba.

ghstack-source-id: ed912f245c5c4ab0894473a184fcea32e281be1f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40199